reconfigured file-structure to convert the application to a package-structure
